Add addBookReview Method Fix Register controller Fix Review model

+++ TubesPBO/src/model/Review.java
@@ -18,7 +18,7 @@ public class Review {
         this.comment = comment;
     }
 
-    public long getIdUser() {
+    public int getIdUser() {
         return idUser;
     }
